requirements-dev.txt
OpenAI
aiohttp
asyncio
configparser
discord
qrcode
pillow
requests
tiktoken
psycopg2
numpy
PyYAML
sentence-transformers
chromadb
fastapi
uvicorn
pydantic
pytest
coverage